refactor(runtimes): extract TimezoneSelect into a shared component
Pulls the timezone dropdown out of runtime-detail.tsx so other surfaces
can reuse the same option list (browser tz first, common IANA zones,
then everything Intl.supportedValuesOf exposes) and the same browser-tz
suffix. TimezoneEditor now just owns the mutation; the picker UI lives
in timezone-select.tsx.
